Merge in BRDM/iocatcher from ~A708881/libfabric-poc:refactor/serialization-step-3-ser-deser-in-lf-connection-final-step-2-fixed to master

* commit '47b20b0f35b874fb23ad110389ade24a2aaaea9a':
  [serialization] Fix PR remarks
  [fix] Fix some compiler warnings
  [serialization] Use the deserialized header
  [serialization] Start to use the serializer in LibfabricConnection

+++ src/server/backends/MemoryBackendBalance.cpp
@@ -115,11 +115,11 @@ void MemoryBackendBalance::deallocate(void * addr, size_t size)
 
 /****************************************************/
 /**
- * Return the memory used by the given memory bacend.
+ * Return the memory used by the given memory backend.
  * Used in unit tests.
  * @param id Id of the memory backend we want to check.
 **/
-size_t MemoryBackendBalance::getMem(int id) const
+size_t MemoryBackendBalance::getMem(size_t id) const
 {
 	//check
 	assert(id >= 0);
